Rename: user.parser 폴더 이동 -> service/user로 이동
