+++ webapp/src/contexts/EssentialForm/useEssentialForm.js
@@ -8,7 +8,7 @@ import authApi from 'api/auth.api';
 
 import { ROUTE } from 'constant/route.constant';
 
-import { essentialInfoParser } from 'service/user.parser';
+import { essentialInfoParser } from 'service/user/user.parser';
 import useUserInfo from 'hooks/useUserInfo';
 import useFileUploader from 'hooks/useFileUploader';
 import userApi from 'api/user.api';
